+++ legend.h
@@ -4,7 +4,7 @@
 
 class legend {
  public:
-  legend(float min, float max, int legend_no = 0, int reversed = 0);
+  legend(double min, double max, int legend_no = 0, int reversed = 0);
   inline int numcolours() { return num_colours; }
   inline int lookup(int idx) {
     if(idx < 0) {
@@ -15,7 +15,7 @@ class legend {
     }
     return colours[idx]; 
   }
-  inline int lookup(float in) {
+  inline int lookup(double in) {
     if(in < min) {
       return colours[0];
     }
@@ -24,17 +24,17 @@ class legend {
     }
     return colours[(int)roundf((in - min) * scale_factor)]; 
   }
-  inline void setmax(float newmax) { max = newmax; adjust_range(); adjust_scale_factor(); }
-  inline void setmin(float newmin) { min = newmin; adjust_range(); adjust_scale_factor(); }
+  inline void setmax(double newmax) { max = newmax; adjust_range(); adjust_scale_factor(); }
+  inline void setmin(double newmin) { min = newmin; adjust_range(); adjust_scale_factor(); }
  private:
   void adjust_scale_factor() { scale_factor = ((num_colours - 1) / range); }
   void adjust_range() { range = max - min; }
   const int* colours;
   int num_colours;
-  float scale_factor;
-  float min;
-  float max;
-  float range;
+  double scale_factor;
+  double min;
+  double max;
+  double range;
 };
 
 #endif
