+++ range.h
@@ -6,13 +6,13 @@
 class range {
  public:
   range() { _modified = 0; }
-  float max() { return _max; }
-  float min() { return _min; }
-  void round(float factor) {
+  double max() { return _max; }
+  double min() { return _min; }
+  void round(double factor) {
     _min = factor * floorf(_min / factor);
     _max = factor * ceilf(_max / factor);
   }
-  void add(float in) { 
+  void add(double in) { 
     if(!_modified) {
       _max = _min = in;
       _modified = 1;
@@ -23,8 +23,8 @@ class range {
   }
  private:
   int _modified;
-  float _max;
-  float _min;
+  double _max;
+  double _min;
 };
 
 #endif
